add bidding progress

+++ site/src/main/java/me/arthurdileo/Transaction.java
@@ -14,10 +14,10 @@ public class Transaction {
 	public String buyer_uuid;
 	public String seller_uuid;
 	public String listing_uuid;
-	public float amount;
+	public double amount;
 	public java.sql.Timestamp created;
 	
-	public Transaction(String buyer_uuid, String seller_uuid, String listing_uuid, float amount) {
+	public Transaction(String buyer_uuid, String seller_uuid, String listing_uuid, double amount) {
 		this.buyer_uuid = buyer_uuid;
 		this.seller_uuid = seller_uuid;
 		this.listing_uuid = listing_uuid;
@@ -25,7 +25,7 @@ public class Transaction {
 	}
 	
 	public Transaction(ResultSet rs) throws SQLException {
-		this(rs.getString("buyer_uuid"), rs.getString("seller_uuid"), rs.getString("listing_uuid"), rs.getFloat("amount"));
+		this(rs.getString("buyer_uuid"), rs.getString("seller_uuid"), rs.getString("listing_uuid"), rs.getDouble("amount"));
 		this.created = rs.getTimestamp("created");
 	}
